Display product of category and sub category.
